split filter bar
