+++ front/src/components/business/FilterGroup/FilterGroup.stories.tsx
@@ -1,7 +1,7 @@
 import { Meta } from '@storybook/react';
 
 import { FilterHistoryItemProps } from '../FilterHistory';
-import { FilterBarDesktop } from './FilterBarDesktop';
+import { FilterBarActuality } from './FilterBarActuality';
 import { FilterGroup } from './FilterGroup';
 import { FilterItem } from './FilterItem';
 import { FilterBarProps } from './types';
@@ -29,8 +29,8 @@ const FILTER_BAR_DATA: FilterBarProps = {
   setOpen: () => console.log('open'),
 };
 
-export const FilterBarExample = (): JSX.Element => (
-  <FilterBarDesktop
+export const FilterBarActualityExample = (): JSX.Element => (
+  <FilterBarActuality
     title={FILTER_BAR_DATA.title}
     searchType={FILTER_BAR_DATA.searchType}
     filterHistories={FILTER_BAR_DATA.filterHistories}
